增加message

+++ client/connecttest/kcpTest/client/client.go
@@ -31,7 +31,7 @@ func main() {
 	str := string(bytes)
 	log.Debug("str is %v", str)
 	//bytes := []byte("test")
-	objM := mnet.NewMsgPackage(uint32(0),bytes, int32(datalen))
+	objM := mnet.NewMsgPackage(uint32(0),bytes, datalen)
 	msg, err := dp.Pack(objM)
 	if err != nil {
 		panic(err)
